+++ Source/ilyvion.LoadingProgress/Settings.cs
@@ -11,11 +11,19 @@ public class Settings : ModSettings
         set => _patchInitialization = value;
     }
 
+    private LoadingWindowPlacement _loadingWindowPlacement = LoadingWindowPlacement.Top;
+    public LoadingWindowPlacement LoadingWindowPlacement
+    {
+        get => _loadingWindowPlacement;
+        set => _loadingWindowPlacement = value;
+    }
+
     public override void ExposeData()
     {
         base.ExposeData();
 
         Scribe_Values.Look(ref _patchInitialization, "patchInitialization", true);
+        Scribe_Values.Look(ref _loadingWindowPlacement, "loadingWindowPlacement", LoadingWindowPlacement.Middle);
     }
 
     public void DoSettingsWindowContents(Rect inRect)
@@ -28,6 +36,38 @@ public class Settings : ModSettings
             ref _patchInitialization,
             "LoadingProgress.PatchInitialization.Tip".Translate());
 
+        if (listingStandard.ButtonTextLabeledPct(
+            "LoadingProgress.LoadingWindowPlacement".Translate(),
+            $"LoadingProgress.{_loadingWindowPlacement}".Translate(),
+            0.6f,
+            TextAnchor.MiddleLeft))
+        {
+            List<FloatMenuOption> list =
+            [
+                new FloatMenuOption(
+                    "LoadingProgress.Top".Translate(),
+                    () => _loadingWindowPlacement = LoadingWindowPlacement.Top),
+                new FloatMenuOption(
+                    "LoadingProgress.Middle".Translate(),
+                    () => _loadingWindowPlacement = LoadingWindowPlacement.Middle),
+                new FloatMenuOption(
+                    "LoadingProgress.Bottom".Translate(),
+                    () => _loadingWindowPlacement = LoadingWindowPlacement.Bottom),
+                // new FloatMenuOption(
+                //     "LoadingProgress.Custom".Translate(),
+                //     () => _loadingWindowPlacement = LoadingWindowPlacement.Custom)
+            ];
+            Find.WindowStack.Add(new FloatMenu(list));
+        }
+
         listingStandard.End();
     }
+}
+
+public enum LoadingWindowPlacement
+{
+    Top,
+    Middle,
+    Bottom,
+    Custom
 }